backend/requirements.txt
# FastAPI 相關套件
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 資料庫相關套件
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# LDAP 相關套件
# python-ldap==3.4.3 需要 Microsoft Visual C++ 14.0 或更高版本
# 安裝方法 1: 先安裝 Microsoft C++ Build Tools: https://visualstudio.microsoft.com/visual-cpp-build-tools/
# 安裝方法 2: 使用預編譯的輪子 (推薦): pip install python_ldap-3.4.3-cp39-cp39-win_amd64.whl
# 安裝方法 3: 使用替代方案 ldap3 (純 Python 實現，無需編譯)
ldap3==2.9.1

# 其他工具套件
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
pytz==2023.3
requests==2.31.0
httpx==0.25.0
aiohttp==3.8.5
